+++ src/main/scala/com/atomist/rug/runtime/js/BaseJavaScriptHandlerFinder.scala
@@ -1,7 +1,8 @@
 package com.atomist.rug.runtime.js
 
 import com.atomist.param.{Parameter, Tag}
-import com.atomist.rug.runtime.{AddressableRug, Rug}
+import com.atomist.project.archive.RugResolver
+import com.atomist.rug.runtime.Rug
 import jdk.nashorn.api.scripting.ScriptObjectMirror
 
 /**
@@ -11,9 +12,9 @@ abstract class BaseJavaScriptHandlerFinder[T <: Rug]
   extends JavaScriptUtils
     with JavaScriptRugFinder[T] {
 
-  override def find(jsc: JavaScriptContext, externalContext: Seq[AddressableRug] = Nil): Seq[T] = {
+  override def find(jsc: JavaScriptContext, resolver: Option[RugResolver] = None): Seq[T] = {
     jsc.vars.collect {
-      case Var(_, handler) if isValidHandler(handler) && handler.getMember("__kind") == kind => extractHandler(jsc, handler, externalContext)
+      case Var(_, handler) if isValidHandler(handler) && handler.getMember("__kind") == kind => extractHandler(jsc,handler)
     }.flatten
   }
 
@@ -31,7 +32,7 @@ abstract class BaseJavaScriptHandlerFinder[T <: Rug]
     obj.hasMember("__kind") && obj.hasMember("__name") && obj.hasMember("__description") && obj.hasMember("handle")
   }
 
-  protected def extractHandler(jsc: JavaScriptContext, handler: ScriptObjectMirror, externalContext: Seq[AddressableRug]): Option[T]
+  protected def extractHandler(jsc: JavaScriptContext, handler: ScriptObjectMirror): Option[T]
 
   protected def name(obj: ScriptObjectMirror): String = obj.getMember("__name").asInstanceOf[String]