+++ src/main/scala/com/atomist/rug/runtime/js/JavaScriptProjectEditor.scala
@@ -1,10 +1,9 @@
 package com.atomist.rug.runtime.js
 
 import com.atomist.param.ParameterValues
-import com.atomist.project.archive.DefaultAtomistConfig
-import com.atomist.project.edit.{ProjectEditorSupport, _}
+import com.atomist.project.archive.{DefaultAtomistConfig, RugResolver}
+import com.atomist.project.edit._
 import com.atomist.rug.kind.core.ProjectMutableView
-import com.atomist.rug.runtime.AddressableRug
 import com.atomist.rug.spi.InstantEditorFailureException
 import com.atomist.source.ArtifactSource
 import com.atomist.util.Timing._
@@ -22,8 +21,8 @@ class JavaScriptProjectEditorFinder
     JsRugOperationSignature(Set("edit"),Set("name", "description")),
     JsRugOperationSignature(Set("edit"), Set("__name", "__description")))
 
-  override def createProjectOperation(jsc: JavaScriptContext, fnVar: ScriptObjectMirror, externalContext: Seq[AddressableRug]): JavaScriptProjectEditor = {
-    new JavaScriptProjectEditor(jsc, fnVar, jsc.rugAs, externalContext)
+  override def createProjectOperation(jsc: JavaScriptContext, fnVar: ScriptObjectMirror, resolver: Option[RugResolver]): JavaScriptProjectEditor = {
+    new JavaScriptProjectEditor(jsc, fnVar, jsc.rugAs, resolver)
   }
 }
 
@@ -34,9 +33,8 @@ class JavaScriptProjectEditor(
                                jsc: JavaScriptContext,
                                jsVar: ScriptObjectMirror,
                                rugAs: ArtifactSource,
-                               externalContext: Seq[AddressableRug]
-                                     )
-  extends JavaScriptProjectOperation(jsc, jsVar, rugAs, externalContext)
+                               resolver: Option[RugResolver])
+  extends JavaScriptProjectOperation(jsc, jsVar, rugAs)
     with ProjectEditorSupport {
 
   override def applicability(as: ArtifactSource): Applicability = Applicability.OK
@@ -51,7 +49,8 @@ class JavaScriptProjectEditor(
       val pmv = new ProjectMutableView(rugAs,
         targetProject,
         atomistConfig = DefaultAtomistConfig,
-        Some(this))
+        Some(this),
+        rugResolver = resolver)
 
       try {
         // Important that we don't invoke edit on the prototype as otherwise all constructor effects are lost!