+++ src/main/scala/com/atomist/rug/runtime/js/JavaScriptRugFinder.scala
@@ -1,10 +1,11 @@
 package com.atomist.rug.runtime.js
 
-import com.atomist.rug.runtime.{AddressableRug, Rug}
+import com.atomist.project.archive.RugResolver
+import com.atomist.rug.runtime.Rug
 
 /**
   * Interface for finding rugs of different types in Nashorn
   */
 trait JavaScriptRugFinder[R <: Rug] {
-  def find(jsc: JavaScriptContext, otherRugs: Seq[AddressableRug]) : Seq[R]
+  def find(jsc: JavaScriptContext, resolver: Option[RugResolver]) : Seq[R]
 }