+++ src/test/scala/com/atomist/rug/runtime/js/JavaScriptEventHandlerTest.scala
@@ -2,7 +2,8 @@ package com.atomist.rug.runtime.js
 
 import com.atomist.graph.GraphNode
 import com.atomist.param.{SimpleParameterValue, SimpleParameterValues}
-import com.atomist.project.archive.{AtomistConfig, DefaultAtomistConfig, RugArchiveReader}
+import com.atomist.project.archive.{AtomistConfig, DefaultAtomistConfig}
+import com.atomist.rug.RugArchiveReader
 import com.atomist.rug.runtime.SystemEvent
 import com.atomist.rug.runtime.plans.{LocalInstructionRunner, LocalPlanRunner}
 import com.atomist.rug.spi.Handlers._
@@ -219,10 +220,10 @@ class JavaScriptEventHandlerTest extends FlatSpec with Matchers with DiagrammedA
     val ts = ClassPathArtifactSource.toArtifactSource("com/atomist/project/archive/MyHandlers.ts")
     val moved = ts.withPathAbove(".atomist/handlers")
     val as = TypeScriptBuilder.compileWithModel(moved)
-    val ops = RugArchiveReader.find(as, Nil)
+    val ops = RugArchiveReader(as)
     val handler = ops.commandHandlers.find(p => p.name == "LicenseAdder").get
     val plan = handler.handle(LocalRugContext(TestTreeMaterializer), SimpleParameterValues(SimpleParameterValue("license", "agpl")))
-    val runner = new LocalPlanRunner(null, new LocalInstructionRunner(Nil, null, null, null))
+    val runner = new LocalPlanRunner(null, new LocalInstructionRunner(handler, null, null, null))
     val response = Await.result(runner.run(plan.get, None), 120.seconds)
     assert(response.log.size === 1)
     response.log.foreach {