+++ src/test/scala/com/atomist/rug/test/gherkin/handler/event/GherkinRunnerEventHandlerTest.scala
@@ -1,6 +1,7 @@
 package com.atomist.rug.test.gherkin.handler.event
 
-import com.atomist.project.archive.{AtomistConfig, DefaultAtomistConfig, RugArchiveReader}
+import com.atomist.project.archive.{AtomistConfig, DefaultAtomistConfig}
+import com.atomist.rug.RugArchiveReader
 import com.atomist.rug.TestUtils._
 import com.atomist.rug.runtime.js.JavaScriptContext
 import com.atomist.rug.test.gherkin._
@@ -34,7 +35,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     val cas = TypeScriptBuilder.compileWithModel(as)
     val msl = new MatchSavingListener
 
-    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader.find(cas)), Seq(msl))
+    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader(cas)), Seq(msl))
     val run = grt.execute()
     run.result match {
       case Passed =>
@@ -58,7 +59,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     //println(ArtifactSourceUtils.prettyListFiles(as))
     val cas = TypeScriptBuilder.compileWithModel(as)
 
-    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader.find(cas)))
+    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader(cas)))
     val run = grt.execute()
     run.result match {
       case Failed(msg) =>
@@ -80,7 +81,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     //println(ArtifactSourceUtils.prettyListFiles(as))
     val cas = TypeScriptBuilder.compileWithModel(as)
 
-    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader.find(cas)))
+    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader(cas)))
     val run = grt.execute()
     run.result match {
       case Failed(msg) =>
@@ -102,7 +103,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     //println(ArtifactSourceUtils.prettyListFiles(as))
     val cas = TypeScriptBuilder.compileWithModel(as)
 
-    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader.find(cas)))
+    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader(cas)))
     val run = grt.execute()
     run.result match {
       case Failed(msg) =>
@@ -131,7 +132,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     //println(ArtifactSourceUtils.prettyListFiles(as))
     val cas = TypeScriptBuilder.compileWithExtendedModel(as)
 
-    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader.find(cas)))
+    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader(cas)))
     val run = grt.execute()
     run.result match {
       case Passed =>
@@ -166,7 +167,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     val msl = new MatchSavingListener
     val grt = new GherkinRunner(
       new JavaScriptContext(cas),
-      Some(RugArchiveReader.find(cas)),
+      Some(RugArchiveReader(cas)),
       listeners = Seq(msl))
     val run = grt.execute()
     assert(msl.matches.nonEmpty)
@@ -187,7 +188,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     val handlerFile = requiredFileInPackage(this, "EventHandlers.ts", atomistConfig.handlersRoot + "/event")
     val as = SimpleFileBasedArtifactSource(Feature1File, passingFeature1StepsFile, handlerFile, nodesFile)
     val cas = TypeScriptBuilder.compileWithModel(as)
-    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader.find(cas)))
+    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader(cas)))
     val run = grt.execute()
     run.result match {
       case Passed =>
@@ -208,7 +209,7 @@ class GherkinRunnerEventHandlerTest extends FlatSpec with Matchers {
     val handlerFile = requiredFileInPackage(this, "EventHandlers.ts", atomistConfig.handlersRoot + "/event")
     val as = SimpleFileBasedArtifactSource(Feature2File, passingFeature1StepsFile, handlerFile, nodesFile)
     val cas = TypeScriptBuilder.compileWithModel(as)
-    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader.find(cas)))
+    val grt = new GherkinRunner(new JavaScriptContext(cas), Some(RugArchiveReader(cas)))
     val run = grt.execute()
     run.result match {
       case Passed =>